requirements.txt
backcall
cachetools
certifi
chardet
cycler
decorator
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
idna
ipykernel
ipython
ipython-genutils
jedi
jupyter-client
jupyter-core
kiwisolver
matplotlib
numexpr
numpy
oauthlib
pandas
parso
pexpect
pickleshare
Pillow
prompt-toolkit
protobuf
ptyprocess
pyasn1
pyasn1-modules
Pygments
pyparsing
python-dateutil
pytz
pyzmq
requests
requests-oauthlib
rsa
six
tables
tornado
traitlets
uritemplate
urllib3
wcwidth